# packages/api-client/src/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/merchants/{merchantId}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get Merchant Details",
        "description": "Retrieves details for a specific merchant by their Merchant ID. Regular admins can only access merchants they created, while superadmins can view any merchant.",
        "parameters": [
          {
            "name": "merchantId",
            "required": true,
            "in": "path",
            "description": "The unique VARCHAR(8) ID of the merchant",
            "schema": {
              "example": "sticpay",
              "type": "string"
            }
          },
          {
            "name": "viewAsAdminId",
            "required": false,
            "in": "query",
            "description": "(SuperAdmin only) View as this Admin ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Merchant details retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantDetailResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden - You did not create this merchant and are not a superadmin."
          },
          "404": {
            "description": "Not Found. Merchant with the specified ID does not exist."
          }
        },
        "tags": [
          "Admin - Merchants"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "update",
        "summary": "Update Merchant",
        "description": "Updates details for a specific merchant.",
        "parameters": [
          {
            "name": "merchantId",
            "required": true,
            "in": "path",
            "description": "The unique VARCHAR(8) ID of the merchant to update",
            "schema": {
              "example": "sticpay",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMerchantDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merchant updated successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MerchantDetailResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request. Invalid update data."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden - You did not create this merchant and are not a superadmin."
          },
          "404": {
            "description": "Not Found. Merchant not found."
          }
        },
        "tags": [
          "Admin - Merchants"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete Merchant (Soft)",
        "description": "Soft deletes a merchant and deactivates associated user/agents.",
        "parameters": [
          {
            "name": "merchantId",
            "required": true,
            "in": "path",
            "description": "The unique VARCHAR(8) ID of the merchant to delete",
            "schema": {
              "example": "oldmerc",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Merchant soft deleted successfully."
          },
          "401": {
            "description": "Unauthorized."
          },
          "403": {
            "description": "Forbidden - You did not create this merchant and are not a superadmin."
          },
          "404": {
            "description": "Not Found. Merchant not found."
          }
        },
        "tags": [
          "Admin - Merchants"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      }
    },
...
    "/admin/merchant-groups": {
      "post": {
        "operationId": "create",
        "summary": "Create a new merchant group",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMerchantGroupsCreateDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The merchant group has been successfully created.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMerchantGroupsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Merchant Groups"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      },
      "get": {
        "operationId": "findAll",
        "summary": "Get all merchant groups",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "skip",
            "required": false,
            "in": "query",
            "description": "Skip specified number of items (alternative to page)",
            "schema": {
              "minimum": 0,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "Search term for group name",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "includeDeleted",
            "required": false,
            "in": "query",
            "description": "Include deleted groups in the results",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "onlyDeleted",
            "required": false,
            "in": "query",
            "description": "Show only deleted groups",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "viewAsAdminId",
            "required": false,
            "in": "query",
            "description": "View as specific admin ID (super admin only)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of merchant groups retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminMerchantGroupsResponseDto"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Merchant Groups"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      }
    },
    "/admin/merchant-groups/{id}": {
      "get": {
        "operationId": "findOne",
        "summary": "Get a merchant group by ID",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The merchant group has been successfully retrieved.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminMerchantGroupsResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Admin - Merchant Groups"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "remove",
        "summary": "Delete a merchant group",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The merchant group has been successfully deleted."
          }
        },
        "tags": [
          "Admin - Merchant Groups"
        ],
        "security": [
          {
            "jwt-bearer-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "EZPG API",
    "description": "API documentation for the EZPG Payment Gateway",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-bearer-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token (Access Token or Temp Token for TFA)",
        "in": "header"
      },
      "jwt-refresh": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "name": "Authorization",
        "description": "Enter Refresh Token",
        "in": "header"
      },
      "csrf": {
        "type": "apiKey",
        "name": "X-CSRF-Token",
        "in": "header",
        "description": "CSRF Token obtained from /auth/csrf-token endpoint"
      }
    },
    "schemas": {
...
      "MerchantDetailResponseDto": {
        "type": "object",
        "properties": {
          "number": {
            "type": "number",
            "description": "Row number"
          },
          "merchantId": {
            "type": "string",
            "description": "Merchant ID"
          },
          "affiliate": {
            "type": "string",
            "description": "Merchant affiliate name"
          },
          "companyName": {
            "type": "string",
            "description": "Merchant company name"
          },
          "groupId": {
            "type": "number",
            "description": "Group ID"
          },
          "groupName": {
            "type": "string",
            "description": "Group name"
          },
          "numberOfMembers": {
            "type": "number",
            "description": "Number of members"
          },
          "status": {
            "type": "string",
            "description": "Merchant status",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ]
          },
          "isActive": {
            "type": "boolean",
            "description": "Whether the merchant is active"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Merchant creation date"
          },
          "createdBy": {
            "type": "string",
            "description": "Admin who created the merchant"
          },
          "depositFeeRate": {
            "type": "number",
            "description": "Deposit fee rate percentage"
          },
          "depositFee": {
            "type": "number",
            "description": "Deposit fee minimum amount (per transaction)"
          },
          "remittanceFeeRate": {
            "type": "number",
            "description": "Remittance fee rate percentage"
          },
          "remittanceFee": {
            "type": "number",
            "description": "Remittance fee minimum amount (per transaction)"
          },
          "foreignCurrencyRemittanceFeeRate": {
            "type": "number",
            "description": "Foreign currency remittance fee rate"
          },
          "foreignCurrencyFeeRate": {
            "type": "number",
            "description": "Foreign currency deposit fee rate"
          },
          "settlementFeeRate": {
            "type": "number",
            "description": "Settlement fee rate percentage"
          },
          "settlementFee": {
            "type": "number",
            "description": "Settlement fee minimum amount (per transaction)"
          },
          "reserveRate": {
            "type": "number",
            "description": "Reserve rate percentage"
          },
          "balance": {
            "type": "number",
            "description": "Current merchant balance"
          },
          "selectedBanks": {
            "description": "Selected banks",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "withdrawalBankName": {
            "type": "string",
            "description": "Withdrawal bank name"
          },
          "withdrawalAccountNumber": {
            "type": "string",
            "description": "Withdrawal account number"
          },
          "withdrawalAccountHolder": {
            "type": "string",
            "description": "Withdrawal account holder"
          },
          "foreignBankName": {
            "type": "string",
            "description": "Foreign bank name"
          },
          "foreignBankAccountNumber": {
            "type": "string",
            "description": "Foreign bank account number"
          },
          "foreignBankAccountHolder": {
            "type": "string",
            "description": "Foreign bank account holder"
          },
          "dashboardId": {
            "type": "string",
            "description": "Dashboard ID"
          },
          "dashboardPassword": {
            "type": "string",
            "description": "Dashboard password (masked)"
          },
          "virtualAccountUsage": {
            "type": "string",
            "description": "Virtual account usage flag (Y/N)"
          },
          "virtualAccountLimit": {
            "type": "number",
            "description": "Virtual account limit"
          },
          "totalDeposit": {
            "type": "number",
            "description": "Total deposit amount"
          },
          "totalDepositFee": {
            "type": "number",
            "description": "Total deposit fee amount"
          },
          "totalWithdrawal": {
            "type": "number",
            "description": "Total withdrawal amount"
          },
          "totalWithdrawalFee": {
            "type": "number",
            "description": "Total withdrawal fee amount"
          },
          "agents": {
            "description": "Top 5 agents with distribution rates",
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "mid": {
            "type": "string",
            "description": "Merchant MID"
          },
          "mkey": {
            "type": "string",
            "description": "Merchant MKEY"
          },
          "callbackUrl": {
            "type": "string",
            "description": "Callback URL"
          },
          "telegramId": {
            "type": "string",
            "description": "Telegram ID"
          },
          "notificationTypes": {
            "description": "Notification types",
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "notificationTime": {
            "type": "string",
            "description": "Notification time"
          }
        },
        "required": [
          "merchantId",
          "affiliate",
          "companyName",
          "groupId",
          "groupName",
          "numberOfMembers",
          "status",
          "isActive",
          "createdAt",
          "createdBy",
          "depositFeeRate",
          "depositFee",
          "remittanceFeeRate",
          "remittanceFee",
          "foreignCurrencyRemittanceFeeRate",
          "foreignCurrencyFeeRate",
          "settlementFeeRate",
          "settlementFee",
          "reserveRate",
          "balance",
          "selectedBanks",
          "withdrawalBankName",
          "withdrawalAccountNumber",
          "withdrawalAccountHolder",
          "foreignBankName",
          "foreignBankAccountNumber",
          "foreignBankAccountHolder",
          "dashboardId",
          "dashboardPassword",
          "virtualAccountUsage",
          "virtualAccountLimit",
          "totalDeposit",
          "totalDepositFee",
          "totalWithdrawal",
          "totalWithdrawalFee",
          "agents",
          "mid",
          "mkey",
          "callbackUrl",
          "telegramId",
          "notificationTypes",
          "notificationTime"
        ]
      },
...
      "UpdateMerchantDto": {
        "type": "object",
        "properties": {
          "affiliate": {
            "type": "string",
            "description": "Updated affiliate (was merchant name)",
            "example": "Updated Affiliate Inc."
          },
          "groupId": {
            "type": "number",
            "description": "Updated ID of the merchant group",
            "example": 2
          },
          "depositFeeRate": {
            "type": "number",
            "description": "Updated deposit fee rate (0.00 - 100.00)",
            "example": 0.6
          },
          "remittanceFeeRate": {
            "type": "number",
            "description": "Updated remittance fee rate (0.00 - 100.00)",
            "example": 1.1
          },
          "reserveRate": {
            "type": "number",
            "description": "Updated reserve rate (0.00 - 100.00)",
            "example": 0.2
          },
          "foreignCurrencyFeeRate": {
            "type": "number",
            "description": "Updated foreign currency fee rate (0.00 - 100.00)",
            "example": 0.5
          },
          "status": {
            "type": "string",
            "description": "Updated status of the merchant",
            "enum": [
              "ACTIVE",
              "INACTIVE",
              "SUSPENDED"
            ],
            "example": "INACTIVE"
          },
          "company_name": {
            "type": "string",
            "description": "Company name of the merchant"
          },
          "telegram_id": {
            "type": "string",
            "description": "Telegram ID of the merchant"
          },
          "min_withdrawal_amount": {
            "type": "number",
            "description": "Minimum withdrawal amount"
          },
          "max_withdrawal_amount": {
            "type": "number",
            "description": "Maximum withdrawal amount"
          },
          "callback_ip": {
            "type": "string",
            "description": "Callback IP address"
          },
          "linked_banks": {
            "type": "string",
            "description": "Linked banks (comma-separated bank codes)",
            "example": "KB001,KB002,KB003"
          },
          "group": {
            "type": "string",
            "description": "Group name"
          }
        }
      },
...
      "AdminMerchantGroupsCreateDto": {
        "type": "object",
        "properties": {
          "groupName": {
            "type": "string",
            "description": "Name of the merchant group",
            "example": "Premium Merchants"
          },
          "status": {
            "type": "string",
            "description": "Status of the merchant group",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          }
        },
        "required": [
          "groupName",
          "status"
        ]
      },
      "AdminMerchantGroupsResponseDto": {
        "type": "object",
        "properties": {
          "groupId": {
            "type": "number",
            "description": "Unique identifier of the merchant group",
            "example": 1
          },
          "groupName": {
            "type": "string",
            "description": "Name of the merchant group",
            "example": "재관리 그룹"
          },
          "status": {
            "type": "string",
            "description": "Current status of the merchant group",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ],
            "example": "ACTIVE"
          },
          "creatorUsername": {
            "type": "string",
            "description": "Username of the admin who created the group",
            "example": "ezpgadmin"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Creation timestamp",
            "example": "2025-05-20T00:00:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update timestamp",
            "example": "2025-05-20T00:00:00.000Z"
          }
        },
        "required": [
          "groupId",
          "groupName",
          "status",
          "creatorUsername",
          "createdAt",
          "updatedAt"
        ]
      },
...
    }
  },
  "security": [
    {
      "csrf": []
    }
  ]
}